Enable editing songs in the current queue
An additional option has been added to the context menu that is used to remove songs from the current queue and becomes visible only when you're right-clicking songs in the current queue. Apart from that, a global variable has been added to keep track of whether you're looking at the current queue playlist.

Without this variable, when you switch to any tab other than the queue tab and right-click a song in the current queue, the extra option does not appear. Also, when you remove a song from the current queue, a number of factors might prevent the playlist from being refreshed visually. This variable helps immediately refresh the playlist to reflect the changed current queue.