Fix #82: Game freezes when playing with custom noteskin without noteskin data. (#115)
ObjectPools were being prefill with invalid noteskin ID before noteskin sanity was being checked, resulting in a null error when a note is being placed back into the pool.

This also fixes massive lag at the start of a song due to creating 1,152 GameNotes per noteskin (12,672 total) to prefill the pool due to it's misplacement within the ObjectPool generation code. It now only generates 144 GameNotes for just the active and valid noteskin.