Keep language strings up-to-date (#117)
The ComboBoxes used to display the language strings for different note colours use an array containing selection menu strings that isn't updated when the language is changed. This is also the case for choosing the landing page. For both situations, the array is created immediately before the ComboBoxes are created, ensuring that the arrays are up-to-date.