Reduce login loading time
Before and after logging in, the game loads 5 kinds of data, 3 of which depends on the global session token. This session token is updated with the user's session token after they have successfully logged in, so these 3 types of data have to be reloaded. However, the other 2 types of data are also reloaded when there is no need to do so.

After logging in, the code now only loads data that depend on the global session token and does not load the ones that don't, which should reduce loading times a bit.