Massive MenuSongSelection refactor.
Refactor MenuSongSelection.as:
A full passthough of all functions, UI elements and event listeners. This reorganizes the entire file and adds full documentation to all functions within the class.

Adds:
	- PageBox: An internal class for display the Page Buttons for the scroll page.
	- HoverPABox: An internal class to add hover ability for highscore entries to display the PA result for the scores.

Fixes:
	- Menu State Saving:
		By using a static classed object, the menu state is kept exactly as it is between switch views, so when returning from the results screen, everything is exactly as it was.

	- Highlighted Song Item:
		The highlighted song item will now correctly update when a song item is moved or removed in response to filters or genre switching.

	- Song Item Mouse Children:
		Was being inaccurately set as enabled for any locked item, should only be enabled for token items .

Other Effected Files:
	- MenuTokens.as: Update MenuSongSelection options reference.
	- PopupFilterManager.as Update MenuSongSelection method name.
	- PopupQueueManager.as Update MenuSongSelection method name.

--------------

Create Flags.as:
This is used to keep track of the temporary flags set throughout the engine and contain the keys within a static location.
It is included in this refactor as MenuSongSelection used a lot of these to keep track of the menu state, but no longer does.

This change affects:
- GlobalVariables.as
- GameResults.as
- MainMenu.as

--------------

Other:
- Optimize ScrollPane.as and ScrollPaneContent.as to simply clear the children instead of create a new sprite each time.