Rewrite UI Box Components
This replaces "Box.as", "BoxText.as", and "BoxButton.as" with more optimized versions.
This eliminates several sprites from each box and clears up several event listeners in the process.
The goal was to replace these classes with visually identical versions.

BoxCheck.as:
The replaces the checkBox.as asset that wasn't easily editable and now follows the same code as the rewritten Box classes.
This fixes check boxes not being considered a button, and they now follow the user color options.
A small white square is also displayed when checked to make it visually easier if a check box is actually checked

LifeBar.as:
This gameplay component replaces the old lifemeterDisplay asset with more optimized version.
While it no longer has the wavy effect, it also contributed to a lot of extra rendering and lag due to its moving nature.

Other:
- The options tabs now use GameBackgroundColor.BG_POPUP instead of the fixed light blue color when active.
- Fixes a unrelated bug where the game background would not display for song loading if on the results screen and changing a setting (etc rates) causes a song reload.
- Fixes a unrelated crash when holding a key in the gameplay editor and exiting.
- Fixes a unrelated bug when music could possible continue to play if the gameplay screen was exited in a strange manner.