Fix Bin Replay MS offset inversion.
Score judging worked one way and bin replays worked the other way. A -75ms early would be recorded as a 75ms late.

Removes the boo verify comparision due to it's in-ability to be compared, as the input array can be in any column order but the output boo have a fixed column order that might not match.

This also updates the replay version and adds a format updater when a bin replay is loaded.