Fix Race condition for SQL Connection
Song Options context menu doesn't always show up when first coming from the login/game loading screen because it's connection is async. But after a song result, the it's already connected so it appears normally.
Fix is just adding a event listener to the connection and toggling the visibility of the menu item instead of add/removing it.