Refactor Song Results (#203)
Song Results Refactor:
Main goal of this refactor was to clean up the general UI for the player due to it's cluttered nature from inclusion of multiple features over the years, and to expand the functionality of the information displayed to the user in the form of graphs.

For the UI, a new BoxButton was introduced call the BoxIcon that allows displays just a icon and provides a optional hover text when the user mouses over the button. These are used several times in the new results screen to compact information into a cleaner format, while not removing the information and text explaining what they do.

These BoxIcons also replace buttons that were also just icons throughout the engine to provide nicer looking icons and the hover text like the Song Selection menu song controls.

Result Graphs:
These were mostly rewritten to use a standardize class format for graphs and to move them away from direct inclusion within the GameResults code. Graphs are now cached and stored instead of redrawn when toggling the graph types, and changing the results for queues to reduce unnecessary redrawing.

Accuracy Graph:
- Rewritten from scratch to be faster.
- Includes text displaying what is early and late to the player, if they are no familiar with the layout of the graph.
- Supports mouse hovering, which displays the note number, accuracy in ms, and the judge given.
- Adds a BoxIcon that when hovered displays the Accuracy and Deviation in both Frames (old way) and MS. This information used to be hidden behind a context-menu on the mods string which was very hard to discover and understand.
- Adds a BoxIcon that toggles the axis of the accuracy graph. This was previously included in the other options in the settings Popup but was had to discover.

Combo Graph:
    The combo graph was update to match the new class format used for graphs.
    Functionality wasn't changed.

GameScoreResult.as:
A game result class that store all values of a gameplay for use on the result screen. This before was just a untyped object which kept management of it's variables a chore and didn't give much flexibility on what the raw data could do.

This class includes helper functions for various result functions that used to be derived on the result screen code.
- Screenshot Path
- Replay Generation
- Accuracy / Deviation in Frames and MS.
- Judge Region Management for Graphs.

BoxIcon.as:
- Simple button class that supports a static Icon and optional hover text.

Other:
- zRanking renamed to SkillRating.
- BoxButton will change it's alpha based on it's disabled state.
- Alert and MouseTooltip have mouse interactive disabled.
- Text will only redraw and update if the actual text value changes.
- MainMenu now uses BoxIcon for the song control buttons instead of drawing them directly.
- MenuSongSelection now uses icons for the genre arrows instead of drawing them directly.
- TimeUtil will display the current time using toLocaleTimeString instead of working out AM/PM itself.
- Binary replays are updated to v1.1 to fix a judge accuracy value inversion bug, a fix for song rate float comparison issues, and temporary removal of the boo comparison check.