Rewrite Noteskins to use Embedded Data instead of network.
Loading the noteskin swf data from the server used to be needed to provided easer ways for noteskins to be added without requiring a new engine build.

Since the inclusion of Custom Noteskins, no new noteskins have been added to the network noteskins for many years and just adds to the engine load time.

This commit embeds all current noteskin data into the game to remove network loading, optimizes how it's loaded, and add documentation and general code cleanup to the overall class.

This also changes how noteskins are stored to remove the need for constent and slow null checks and fallbacks by filling any possible gaps in the noteskin data during the loading and verifying stage.