Fix for direct WebSocket connections.
Accessing localhost:<port_number> would crash the game as it only expects websocket connections and didn't check before attempting to write a response.

Add a fields check that checks for websocket fields and bails if none found.

Also adds a instructions button under the overlay toggle to explain the feature.