Anti-GPU Rampdown
By doing a sparse but steady amount of screen updates using a single pixel in the top left, the GPU is kept active on laptops. This fixes the issue when a skip can appear to happen when the GPU re-awakes to begin drawing updates after a break in a song.