Add Column Filter To Results Screen (#274)
* Implemented Column FIlter

the feature is there, but it looks ugly and the user has no clue what is being filtered

* Icon changing

the icon now changes when you click on the button, except the icons dont exist!

* Better UI

Uses BoxButton instead of BoxIcon. Made it so BoxButton can have hover text. Made it so it uses alpha instead of messing with rgb mumbo jumbo. needs lang strings for the column filters.

* alpha values changed

made the alpha less and default a normal number

* fixed replays

replays (new and old) no longer break the game once the song ends. New replays are still compatible with older engine versions

* modified naming of filters

LD and UR are easy to confuse with many other things. LHS and RHS better represent what it is showing.

* Fixed Replays pt2

Made changes requested by @Psycast

* Removed Unused import

iconSmallT  was not being used but was imported.