Fix MP Spectating (#288)
* Fixed UI Bug

Fixed bug where disconnect button was visible ontop of options button on MP results screen.

* Fixed state related bug

Fixed a bug that caused the player to not enter gameplay when finishing loading a song in mp.

* Fixed loaded state error

fixed a bug where when a player left the room after loading a song the room would keep the slot status as loaded which caused serious loading issues.

* Fixed results display bug

the results are no longer displayed several times when any status change happens.

* Added new button

added a button to control when a game starts and when to spectate the game. This may not work at all and is only visual

* Spectating works!

Players now have the ability to spectate(?) games! The ready button now also works. There is some number manipulation to have older versions work but it works.

* Spectating and editor works

You can now spectate a single person and the multiplayer editor isn't completely busted! The spectating has some issues (no live PA count) but the layout works

* live spectate stuff works

It all works now! i think. Spectating is back and better then ever!(?) PA and Combo etc. works now

* Can spectate playing rooms

You can now start spectating rooms when they are already in game.

* removed todo

not sure what this is for and neither does Zag. Also removed a whitespace.